1.增加ｋｅｙ发送和模拟输入. 2.优化ｔｃｐ接收数据，规定一个　特殊结束标志

+++ XDEService/src/com/xcy/net/ConnListener.java
@@ -8,5 +8,5 @@ public interface ConnListener {
 	public void RemoveClient(String client);
 	
 	public void HandlerKey(int keycode,int action);
-	public void HandlerTouch(float x,float y,int action);
+	public void HandlerTouch(int x,int y,int action);
 }
